kraliki-lab/services/recall-kraliki/backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# GLM 4.6 API (ZhipuAI)
zhipuai>=2.0.0

# Storage
python-frontmatter>=1.0.0
PyYAML>=6.0.0

# Search
whoosh>=2.7.4  # Keyword search
numpy>=1.26.0  # Vector operations

# Utils
python-dotenv>=1.0.0
sniffio>=1.3.0
aiofiles>=23.0.0

# FastMCP (for MCP server)
fastmcp>=0.2.0

# Optional: Database (if needed later)
# sqlalchemy==2.0.25
# alembic==1.13.1
